requirements.txt
# NOTE: Dependencies use version ranges rather than exact pins. Railway uses
# pip install which resolves the latest compatible version within each range
# on every deploy. For fully reproducible builds, run `pip freeze > requirements.lock`
# and reference that file instead. The ranges below are kept intentionally to
# receive security patches automatically between deploys.
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
websockets>=14,<15
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
psycopg2-binary>=2.9,<3.0
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0
PyJWT[crypto]>=2.8,<3.0
bcrypt>=4.0,<5.0
python-multipart>=0.0.7
httpx>=0.27
supabase>=2.0
openai>=1.0
google-auth>=2.0,<3.0
google-auth-oauthlib>=1.0,<2.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
python-telegram-bot>=21,<22
apscheduler>=3.10,<4.0
boto3>=1.35,<2.0
Pillow>=11,<12
pytz>=2024.1
pyotp>=2.9,<3.0
qrcode[pil]>=8,<9
python-magic>=0.4,<1.0
slowapi>=0.1,<1.0
cryptography>=43,<44
pdfplumber>=0.11,<1.0
